Add log-in permission on event's helpers

+++ app/helpers/events_helper.rb
@@ -4,16 +4,20 @@ module EventsHelper
   end
 
   def current_user_attended_event
-    if !current_user.attended_events.find_by_id(@e_id)
-      @go = link_to 'Go', attendances_path(:event_id => @event.id, :user_id => current_user.id), :method => :post, class: 'btn btn-success'
-    else
-      @leave = link_to 'Leave this event', attendance_path(:event_id => @event.id, :user_id => current_user.id), method: :delete, class: 'btn btn-success'
+    if user_sign_in?
+      if !current_user.attended_events.find_by_id(@e_id)
+        @go = link_to 'Go', attendances_path(:event_id => @event.id, :user_id => current_user.id), :method => :post, class: 'btn btn-success'
+      else
+        @leave = link_to 'Leave this event', attendance_path(:event_id => @event.id, :user_id => current_user.id), method: :delete, class: 'btn btn-success'
+      end
     end
   end
 
   def del_btn
-    if @event.user.name == current_user.name
-      @del = link_to 'Delete event', event_path(@event), method: :delete, class: 'btn btn-success'
+    if user_sign_in?
+      if @event.user.name == current_user.name
+        @del = link_to 'Delete event', event_path(@event), method: :delete, class: 'btn btn-success'
+      end
     end
   end
